Synchronize Main Pages
Combining all three main pages into only one now (work properly) to make it suitable for frontend sections
