+++ AccessibilityPortal/src/Pages/MainPage/mainpage.tsx
@@ -1,0 +1,232 @@
+import BasicCardComponent from "../../CommonComponents/Card/BasicCardComponent";
+import '../../Styles/main.scss';
+import '../../Styles/custom.scss';
+import surveyIcon from "../../assets/rate-rating-survey-3-svgrepo-com.svg";
+import profileIcon from "../../assets/profile-circle-svgrepo-com.svg";
+import studiesIcon from "../../assets/book-education-study-svgrepo-com.svg";
+import templateIcon from "../../assets/dashboard-layout-svgrepo-com.svg";
+import BasicButtonComponent from "../../CommonComponents/Buttons/BasicButtonComponent"; 
+import firebase from 'firebase/compat/app';
+import { collection, query, where, getDocs } from "firebase/firestore";
+import React, { useEffect, useState } from 'react';
+import { Link, useNavigate } from "react-router-dom";
+import { Button } from 'reactstrap';
+import ErrorMessage from '../../CommonComponents/ErrorMessage';
+import '../../Styles/login.scss';
+import { Providers, auth, db } from '../../configurations/firebase';
+import logging from '../../configurations/logging';
+
+function MainPage(){
+
+    const navigate = useNavigate();
+
+    const directToChangePassword = () => {
+        navigate('/change');
+    };
+
+    const directToLogoutPage = () => {
+        navigate('/logout');
+    };
+
+    const directToStudyPage = () => {
+        navigate('/studies');
+    }
+
+    const directToSurveyPage = () => {
+        navigate('/survey');
+    }
+
+    const directToProfilePage = () => {
+        navigate('/profile');
+    }
+
+    const directToTemplatePage = () => {
+        navigate('/template');
+    }
+
+    const [isAdmin, setIsAdmin] = useState<boolean>(false);
+    const [isResearcher, setIsResearcher] = useState<boolean>(false);
+    const [isParticipant, setIsParticipant] = useState<boolean>(false);
+
+    useEffect(() => {
+        auth.onAuthStateChanged( async user => {
+            if (user) {
+                if(user.emailVerified) {
+                    logging.info('User detected.' + user.email);
+                    // User.mail to lead to correct main page
+
+                    const q = query(collection(db, "users"), where("email", "==", user.email));
+                    const querySnapshot = await getDocs(q);
+
+                    querySnapshot.forEach((doc) => {
+                        console.log(doc.id, ' => ', doc.data().role);
+                        // Detect ROLE of user
+                        let userRole = doc.data().role;
+                        switch(userRole) {
+                            case 'admin':
+                                setIsAdmin(true)
+                                break
+                            case 'researcher':
+                                setIsResearcher(true)
+                                break
+                            case 'participant':
+                                setIsParticipant(true)
+                                break
+                            default:
+                                return 'unknown';
+                        }
+                    });
+            }}
+    })}, []);
+
+
+    return (
+        <div className="main-page">
+
+            <div className="header">
+            <div className="input-group">
+                <input type="search" className="form-control"></input>
+                <button className="btn btn-dark">Search</button>
+            </div>
+            </div>
+
+        { isAdmin ?
+        <>
+            <div className="display-box">
+                <div>
+                    <BasicCardComponent 
+                        imageUrl={profileIcon}
+                        title={"Users"}
+                    ></BasicCardComponent>
+                </div>
+                <div>
+                    <BasicCardComponent
+                        imageUrl={templateIcon}
+                        title={"Messages"}
+                    ></BasicCardComponent>
+                </div>
+                <div>
+                    <BasicCardComponent
+                        imageUrl={surveyIcon}
+                        title={"Surveys"}
+                    ></BasicCardComponent>
+                </div>
+                <div>
+                    <BasicCardComponent
+                        imageUrl={studiesIcon}
+                        title={"Studies"}
+                    ></BasicCardComponent>
+                </div>
+                <div>
+                    <BasicCardComponent
+                        imageUrl={profileIcon}
+                        title={"Profile"}
+                    ></BasicCardComponent>
+                </div>
+                <div>
+                    <BasicCardComponent
+                        imageUrl={templateIcon}
+                        title={"Templates"}
+                    ></BasicCardComponent>
+                </div>
+
+                <div>
+                    <BasicButtonComponent color='light' title={"Do Something"} onClick={directToChangePassword}></BasicButtonComponent>
+                </div>
+
+                <div>
+                    <BasicButtonComponent color='light' title={"Log Out"} onClick={directToLogoutPage}></BasicButtonComponent>
+                </div>
+                
+            </div>
+        </>
+        : isResearcher ?
+        <>
+            <div className="display-box">
+                <div>
+                    <BasicCardComponent 
+                        imageUrl={surveyIcon}
+                        title={"Surveys"}
+                    ></BasicCardComponent>
+                </div>
+                <div>
+                    <BasicCardComponent
+                        imageUrl={studiesIcon}
+                        title={"Studies"}
+                    ></BasicCardComponent>
+                </div>
+                <div>
+                    <BasicCardComponent
+                        imageUrl={profileIcon}
+                        title={"Profile"}
+                    ></BasicCardComponent>
+                </div>
+                <div>
+                    <BasicCardComponent
+                        imageUrl={templateIcon}
+                        title={"Templates"}
+                    ></BasicCardComponent>
+                </div>
+
+                <div>
+                    <BasicButtonComponent color='light' title={"Change My Password"} onClick={directToChangePassword}></BasicButtonComponent>
+                </div>
+
+                <div>
+                    <BasicButtonComponent color='light' title={"Log Out"} onClick={directToLogoutPage}></BasicButtonComponent>
+                </div>
+                
+            </div>
+        </>
+        : isParticipant ?
+        <>
+            <div className="display-box">
+                <div>
+                    <BasicCardComponent 
+                        imageUrl={surveyIcon}
+                        title={"Surveys"}
+                        handleClick={directToSurveyPage}
+                    ></BasicCardComponent>
+                </div>
+                <div>
+                    <BasicCardComponent
+                        imageUrl={studiesIcon}
+                        title={"Studies"}
+                        handleClick={directToStudyPage}
+                    ></BasicCardComponent>
+                </div>
+                <div>
+                    <BasicCardComponent
+                        imageUrl={profileIcon}
+                        title={"Profile"}
+                        handleClick={directToProfilePage}
+                    ></BasicCardComponent>
+                </div>
+                <div>
+                    <BasicCardComponent
+                        imageUrl={templateIcon}
+                        title={"Templates"}
+                        handleClick={directToTemplatePage}
+                    ></BasicCardComponent>
+                </div>
+
+                <div>
+                    <BasicButtonComponent color='light' title={"Change My Password"} onClick={directToChangePassword}></BasicButtonComponent>
+                </div>
+
+                <div>
+                    <BasicButtonComponent color='light' title={"Log Out"} onClick={directToLogoutPage}></BasicButtonComponent>
+                </div>
+                
+            </div>
+        </>
+        :
+        <>
+
+        </>
+    }
+        </div>
+    );
+}
+
+export default MainPage;
